SearchByIngredient Page update 2.
Error to investigate.

Searching for drinks by ingredient uses the 1st API which returns an array of drinksObject(with only keys strDrink, strDrinkThumb, and idDrink). The page renders each drink component with no issue, albeit lacking ingredients and instructions info to pass down to each drink component's modal.

I attempted to call a 2nd API, that is idDrink specific and supplies the ingredients and instructions info. This 2nd API is called for each drink component when rendered. ISSUE is this 2nd API doesn't stop calling.